"Features without documentation" into temp-master (#21)
* Completely refactor internal structure

* GET /auth/me

* Filter exercises by user_id(s)

* Delete own exercises

* Add test for simple user /bulk/delete_exercises
