+++ api.yml
@@ -440,7 +440,7 @@ paths:
             application/json:
               schema:
                 $ref: "#/components/schemas/ErrorObject"
-  /api/bulk_create_exercises:
+  /api/bulk/create_exercises:
     post:
       summary: "Creates multiple exercises into the system"
       tags:
@@ -499,7 +499,7 @@ paths:
             application/json:
               schema:
                 $ref: "#/components/schemas/ErrorObject"
-  /api/bulk_create_or_find_tag_categories:
+  /api/bulk/create_or_find_tag_categories:
     post:
       summary: "Creates or Find tag categories"
       tags:
@@ -530,7 +530,7 @@ paths:
             application/json:
               schema:
                 $ref: "#/components/schemas/ErrorObject"
-  /api/bulk_modify_exercises_validity:
+  /api/bulk/modify_exercises_validity:
     put:
       summary: "Validate / Unvalidate given exercises"
       operationId: "ValidateExercises"
@@ -572,7 +572,7 @@ paths:
             application/json:
               schema:
                 $ref: "#/components/schemas/ErrorObject"
-  /api/bulk_delete_exercises:
+  /api/bulk/delete_exercises:
     delete:
       summary: "Delete given exercises"
       operationId: "DeleteExercises"
