+++ routes/api/bulk/bulk_delete_exercises.js
@@ -7,6 +7,60 @@ const Sequelize = require("sequelize");
 const Op = Sequelize.Op;
 
 module.exports = function (req, res, next) {
+    return prevent_security_issue(req.user, req.body)
+        .then(() => delete_exercises(req.body))
+        .then(([files, destroyedRowNumber]) => {
+            console.log("Successfully destroy row(s) " + destroyedRowNumber);
+            fileManager
+                .delete_stored_files(files)
+                .then(() => res.status(200).end());
+        })
+        .catch(/* istanbul ignore next */
+            (err) => next(err));
+};
+
+// As this endpoint is available for any type of user, we must prevent deletions of other exercise(s)
+// if we aren't an admin
+function prevent_security_issue({role, id}, exercises_ids) {
+    return new Promise((resolve, reject) => {
+        if (role === "admin") {
+            resolve();
+        } else {
+            models
+                .Exercise
+                .findAll({
+                    attributes: [[Sequelize.fn('DISTINCT', Sequelize.col('user_id')), 'user']],
+                    where: {
+                        id: {
+                            [Op.in]: exercises_ids
+                        }
+                    }
+                }, {
+                    rejectOnEmpty: true
+                })
+                .then((exercises_creators) => {
+                    const all_from_this_user = exercises_creators
+                        .map(creator => creator.user)
+                        .every((creator) => creator === id);
+                    // I don't want to create a complete test case just for that
+                    /* istanbul ignore if */
+                    if (all_from_this_user) {
+                        resolve();
+                    } else {
+                        let error = new Error("FORBIDDEN");
+                        error.message = "It seems you tried to delete somebody else exercise(s) : " +
+                            "This incident will be reported";
+                        error.status = 403;
+                        throw error;
+                    }
+                })
+                .catch(/* istanbul ignore next */
+                    (err) => reject(err));
+        }
+    })
+}
+
+function delete_exercises(exercises_ids) {
     // transaction here as if anything bad happens, we don't commit that to database
     // Here we need the serializable level because we should be able to prevent an user
     // to modify one of these exercises while we deleted them
@@ -25,7 +79,7 @@ module.exports = function (req, res, next) {
                             [Op.ne]: null
                         },
                         id: {
-                            [Op.in]: req.body
+                            [Op.in]: exercises_ids
                         }
                     }
                 }, {
@@ -44,19 +98,11 @@ module.exports = function (req, res, next) {
                                 transaction: t,
                                 where: {
                                     id: {
-                                        [Op.in]: req.body
+                                        [Op.in]: exercises_ids
                                     }
                                 }
                             })
                     ]);
                 });
-        })
-        .then(([files, destroyedRowNumber]) => {
-            console.log("Successfully destroy row(s) " + destroyedRowNumber);
-            fileManager
-                .delete_stored_files(files)
-                .then(() => res.status(200).end());
-        })
-        .catch(/* istanbul ignore next */
-            (err) => next(err));
-};
+        });
+}
