+++ routes/api/tags_categories/getTagCategories.js
@@ -1,0 +1,15 @@
+const models = require('../../../models');
+
+module.exports = (req, res, next) => {
+    return models
+        .Tag_Category
+        .findAll({
+            attributes: [
+                "id",
+                ["kind", "category"]
+            ]
+        })
+        .then((result) => res.json(result))
+        .catch(/* istanbul ignore next */
+            err => next(err))
+};
